refactor(core): replace watcher with store updater cmp
