+++ packages/core/src/container/VueFlow/VueFlow.vue
@@ -12,6 +12,7 @@ import { useHooks } from '../../store/hooks'
 import EdgeRenderer from '../EdgeRenderer/EdgeRenderer.vue'
 import NodeRenderer from '../NodeRenderer/NodeRenderer.vue'
 import { useStylesLoadedWarning } from '../../composables/useStylesLoadedWarning'
+import { StoreUpdater } from './StoreUpdater'
 
 const props = withDefaults(defineProps<FlowProps>(), {
   snapToGrid: undefined,
@@ -116,5 +117,7 @@ export default {
     <slot />
 
     <A11yDescriptions />
+
+    <StoreUpdater :opts="props" />
   </div>
 </template>
